% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaorunARC.R
\name{aaorunARC}
\alias{aaorunARC}
\title{Reduce an AAT 2df+AAOmega arc file using 2dfDR function aaorun}
\usage{
aaorunARC(file = file, idx = idx, doDark = T, darkFile = darkFile,
  doBias = T, biasFile = biasFile, tlmFile = tlmFile)
}
\arguments{
\item{file}{Path to the input fits file to be reduced}

\item{idx}{Path to the idx aaorun paramter file}

\item{doDark}{TRUE/FLASE, do you want to perform dark frame correction}

\item{darkFile}{If doDark=TRUE, path to the input FITS dark frame}

\item{doBias}{TRUE/FLASE, do you want to perform bais frame subtraction}

\item{biasFile}{If doBias=TRUE, path to the input FITS bias frame}

\item{tlmFile}{Path to the input FITS tramline map}
}
\description{
This function perfoms the data reduction on a 2dfDR arc file.
 You must provide an idx filename and tramline file name (potentially produced by aaorunTLM.R). 
 Optionally you can also perform a bias subtraction and dark correction if provided (this is recomended)
}
\examples{
aaorunARC(file='arc2dF.fits', idx='gama_blue.idx', doDark=T,darkFile='DARKmaster.fits',doBias=T, biasFile='BIASmaster.fits', tlmFile='TLM.fits')
}
