% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaorunFLAT.R
\name{aaorunFLAT}
\alias{aaorunFLAT}
\title{Reduce an AAT 2df+AAOmega flat file using 2dfDR function aaorun}
\usage{
aaorunFLAT(file = file, idx = idx, doDark = T, darkFile = darkFile,
  doBias = T, biasFile = biasFile, arcFile = arcFile, waveStart = "NA",
  waveEnd = "NA")
}
\arguments{
\item{file}{Path to the input file to be reduced}

\item{idx}{Path to the idx aaorun paramter file}

\item{doDark}{TRUE/FLASE, do you want to perform dark frame correction}

\item{darkFile}{If doDark=TRUE, path to the input dark frame}

\item{doBias}{TRUE/FLASE, do you want to perform bais frame subtraction}

\item{biasFile}{If doBias=TRUE, path to the input bias frame}

\item{arcFile}{Path to the input arc file}

\item{waveStart}{The lower bound of the wavelemngth range to use in the reduction}

\item{waveEnd}{The upper bound of the wavelemngth range to use in the reduction}
}
\description{
This function perfoms the data reduction on a 2dfDR flat file.
 You must provide an idx filename and arc file name (potentially produced by aaorunARC.R). 
 Optionally you can also perform a bias subtraction and dark correction if provided (this is recomended)
}
\examples{
aaorunFLAT(file='flat2dF.fits', idx='gama_blue.idx', doDark=T,darkFile='DARKmaster.fits',doBias=T, biasFile='BIASmaster.fits', arcFile='arc2dF.fits')
}
