% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaorunObj.R
\name{aaorunObj}
\alias{aaorunObj}
\title{Reduce an AAT 2df+AAOmega object file using 2dfDR function aaorun}
\usage{
aaorunObj(file = file, idx = idx, doDark = T, darkFile = darkFile,
  doBias = T, biasFile = biasFile, flatFile = flatFile,
  tlmFile = tlmFile, arcFile = arcFile)
}
\arguments{
\item{file}{Path to the input file to be reduced}

\item{idx}{Path to the idx aaorun paramter file}

\item{doDark}{TRUE/FLASE, do you want to perform dark frame correction}

\item{darkFile}{If doDark=TRUE, path to the input dark frame}

\item{doBias}{TRUE/FLASE, do you want to perform bais frame subtraction}

\item{biasFile}{If doBias=TRUE, path to the input bias frame}

\item{flatFile}{Path to the input flat file}

\item{tlmFile}{Path to the input tramline map file}

\item{arcFile}{Path to the input arc file}
}
\description{
This function perfoms the data reduction on a 2dfDR object file.
 You must provide an idx filename, flat file (potentially produced by aaorunFLAT.R), 
 arc file name (potentially produced by aaorunARC.R), and tramline map (potentially produced by aaorunTLM.R). 
 Optionally you can also perform a bias subtraction and dark correction if provided (this is recomended)
}
\examples{
aaorunObj(file='object2dF.fits', idx='gama_blue.idx', doDark=T,darkFile='DARKmaster.fits',doBias=T, biasFile='BIASmaster.fits', flatFile='flat2dF.fits',tmlFile='TLM.fits',arcFile='arc2dF.fits')
}
