% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaorunTLM.R
\name{aaorunTLM}
\alias{aaorunTLM}
\title{Reduce an AAT 2df+AAOmega tramline map file using 2dfDR function aaorun}
\usage{
aaorunTLM(file = file, idx = idx, doDark = T, darkFile = darkFile,
  doBias = T, biasFile = biasFile, outname = outname)
}
\arguments{
\item{file}{Path to the input file to be reduced}

\item{idx}{Path to the idx aaorun paramter file}

\item{doDark}{TRUE/FLASE, do you want to perform dark frame correction}

\item{darkFile}{If doDark=TRUE, path to the input dark frame}

\item{doBias}{TRUE/FLASE, do you want to perform bais frame subtraction}

\item{biasFile}{If doBias=TRUE, path to the input bias frame}

\item{outname}{output filename of tramline map}
}
\description{
This function perfoms produces and aaorun tramline map form a given file.
 You must provide an idx filename. 
 Optionally you can also perform a bias subtraction and dark correction if provided (this is recomended)
}
\examples{
aaorunTML(file='arc2dF.fits', idx='gama_blue.idx', doDark=T,darkFile='DARKmaster.fits',doBias=T, biasFile='BIASmaster.fits',outname='TLM.fits')
}
