% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddDEVILS_VISCLASS.R
\name{AddDEVILS_VISCLASS}
\alias{AddDEVILS_VISCLASS}
\title{Updates DEVILS_CurrentVIS and DMCat$DEVILS_UTMJD_VIS to DMCat}
\usage{
AddDEVILS_VISCLASS(DMCatPath = DMCatPath, VisDir = VisDir,
  UTMJD = NA, verbose = T)
}
\arguments{
\item{DMCatPath}{Path to the location of the DMCat you want to update}

\item{VisDir}{The directory path of the visual calisifications (must contain sub-folder 'Good' and 'Bad')}

\item{UTMJD}{UTMJD to add to DEVILS_UTMJD_VIS column. If NA, current date will be added.}

\item{vebose}{Tell me what's going on}
}
\description{
Function takes in ID from a given directory in folders called 'Good' and 'Bad'
everything in 'Good' gets assinged a DEVILS_CurrentVIS==1, everything that is in 'Bad' gets a 
DEVILS_CurrentVIS==0. The UTMJD of the current or a specific date is added to DEVILS_UTMJD_VIS.
}
\examples{
AddDEVILS_VISCLASS(DMCatPath='catalogues/MASTERcats/DMCat2018-12-31.rda', VisDir='/Users/luke/work/DEVILS/VIS_CHECKS_copy/', UTMJD=NA)
}
