% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VacuumFromAir.R
\name{AirFromVacuum}
\alias{AirFromVacuum}
\title{Internal AutoZ routines for converting air wavelengths to vacuum wavelengths}
\usage{
AirFromVacuum(vac)
}
\arguments{
\item{Vac}{Wavelength in vacuum}
}
\value{
Wavelength in air
}
\description{
Function interpolate vacuum wavelengths to air wavelengths
using a lookup table. Equation comes from the SDSS web page:
http://www.sdss.org/dr7/products/spectra/vacwavelength.html. 
 Reference to Morton (1991, ApJS, 77, 119).
}
\examples{
Air<-AirFromVacuum(3728.516)
}
\author{
I. Baldry
}
