% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizCheck.R
\name{bizCheck}
\alias{bizCheck}
\title{Check DEVILS directory sturucture and print warnings for wrong file types}
\usage{
bizCheck(workingDir = ".", stopError = T, logName = logName,
  verbose = verbose)
}
\arguments{
\item{workingDir}{Path to top level of DEVILS directory structure}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}
}
\description{
This function perfoms the data reduction on a 2dfDR arc file.
 You must provide an idx filename and tramline file name (potentially produced by aaorunTLM.R). 
 Optionally you can also perform a bias subtraction and dark correction if provided (this is recomended)
}
\examples{
bizCheck(workingDir='.', logName='tempLog.txt', verbose=1)
}
