% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCal.R
\name{checkCal}
\alias{checkCal}
\title{Produce QC diagnositic plots for dark and bias frames}
\usage{
checkCal(biasDir = biasDir, darksDir = darksDir, verbose = 1)
}
\arguments{
\item{biasDir}{directory containing bias frames}

\item{darksDir}{directory containing dark frames}

\item{verbose}{0, 1,2 , let me know whats going on. 0=no, 1=a bit, 2=a lot}
}
\description{
Produces diagnostic plots of all biases/darks in a given directory. Will
plot and image of the calibration from and a number of diagnositcs.
}
\examples{
checkCal(biasDir='data/biases/run2_2018_01/', darksDir='data/darks/run2_2018_01/', verbose=2)
}
