% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConfig.R
\name{checkConfig}
\alias{checkConfig}
\title{Check Configuration Files are consistent with DMCat}
\usage{
checkConfig(configFiles = configFiles, DMCatN = DMCatN,
  logName = logName, verbose = verbose)
}
\arguments{
\item{configFiles}{Vector of paths to config files to be checked.}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}

\item{DMCat}{Path to current DMCat to compare against}
}
\description{
Checks that final output configuration files are consistent with
the current DMCat in terms of fields, IDs, priorities, star-galaxy flags, VISCLASS flags and
mask flags.
}
