% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckSync.R
\name{CheckSync}
\alias{CheckSync}
\title{Check if OwnCloud Sync is still running}
\usage{
CheckSync(dir = "data/rawdata", pauseTime = 3, verbose = 2,
  logName = "tmp.txt")
}
\arguments{
\item{dir}{Path to directory to check}

\item{pauseTime}{Time to wait between checks in min}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}
}
\description{
Checks to see if a directory is still changing in size. If it is, 
it waits for an allocated amount of time and checks again.
}
\examples{
CheckSync()
}
