% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoutConfig.R
\name{cutoutConfig}
\alias{cutoutConfig}
\title{Produces cutout YJK RGB images for all sources in a fibre configuration}
\usage{
cutoutConfig(configFile = configFile, size = size, outDir = outDir,
  cores = cores)
}
\arguments{
\item{configFile}{Path to configuration file to read}

\item{size}{Size of the desired cutouts in Arcsec}

\item{outDir}{Path for directory to write files to}

\item{cores}{Number of cores to run over}
}
\description{
This is function takes a specified DEVILS fibre configuration
file (.lis) and produces YJK cutout images of all sources in that file. 
This also notes which fibres are placed on targerts, standards, sky, or
guidestars. NOTE: This must be run on Munro!
}
