% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2jd.R
\name{date2jd}
\alias{date2jd}
\title{Covert easily-readable common date to Julian date}
\usage{
date2jd(year = 1968, mon = 5, mday = 23, hour = 12)
}
\arguments{
\item{year}{year to convert}

\item{mon}{month to convert}

\item{mday}{day to convert}

\item{hour}{hour to convert}
}
\value{
Julian Date
}
\description{
Covert the date in year, month, day and hour in Jullian date
}
\examples{
date2jd(year=2017,mon=12,mday=18,hour=12) 
}
