% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{DoCrossCorr}
\alias{DoCrossCorr}
\title{Compute cross corelations un AutoZ}
\usage{
DoCrossCorr(spec, gap, tempData, helioVel, plan, z_prior, highZ = FALSE)
}
\arguments{
\item{spec}{input spectrum containing spec$lambda and spec$flux}

\item{gap}{gap in shifts probed in cross correlation}

\item{plan}{fast-forier tranform plan}

\item{z_prior}{redshift priors, c(lo,hi)}

\item{tempDatar}{structure of template data}

\item{heloval}{helocentric correction value}
}
\description{
Function to compute cross-correlation strngth between a given spectrum and
a set of templates. Function is internal to AutoZ.
}
\examples{
None applicable as internal function.... 
}
