% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementshift.R
\name{elementshift}
\alias{elementshift}
\title{Apply element shift to vector in Hanning Smoothing for plotting}
\usage{
elementshift(x, dx)
}
\arguments{
\item{x}{vector}

\item{dx}{shift required}
}
\description{
Apply element shift to vector in Hanning smoothing function.
}
\examples{
None applicable as internal function.... 
}
