% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNewSpec.R
\name{extractNewSpec}
\alias{extractNewSpec}
\title{Extract 1D spectra from a reduced FITS file}
\usage{
extractNewSpec(file = file, logName = logName, verbose = verbose,
  makePlot = F, zeroPoint = T, NowDate = NowDate)
}
\arguments{
\item{file}{reduced 2dF+AAOmega frame to extract spectra from}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}

\item{makePlot}{make 1D plots of extracted spectra}

\item{zeroPoint}{TRUE/FALSE apply pre-calcuated (from fluxCal.R) zeroPoint flux scaling}

\item{doCosmic}{TRUE/FALSE do additonal cosmic ray rejection}
}
\description{
Function extracts 1D spectra from a reduced 2dF+AAOmega frame and
writes these to the TAZ data structure. Internal function to TAZ
}
\examples{
extractNewSpec(file='objected2df_red.fits', logName='tmpLog.txt', verbose=1, makePlot=F, zeroPoint=F)
}
