% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNewData.R
\name{findNewData}
\alias{findNewData}
\title{Find unreduced raw data in TAZ data structure}
\usage{
findNewData(logName = logName, verbose = verbose)
}
\arguments{
\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}
}
\value{
list of date directories that have not been reduced
}
\description{
Function searches the TAZ raw data folders and identifies dates which
have not currently been reduced
}
\examples{
toReduce<-findNewData(logName='tempLog.txt', verbose=1)
}
