% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxCal.R
\name{fluxCal}
\alias{fluxCal}
\title{Flux calibrate reduced data frames}
\usage{
fluxCal(file = file, stdStars = stdStars, logName = logName,
  verbose = verbose)
}
\arguments{
\item{file}{reduced 2df+AAOmega frame to be calibrated.}

\item{stdStars}{catalogue of standard stars contining ID, 
RA, DEC, MagG, MagR, MagI, MagG_filtername, MagR_filtername, MagI_filtername (see getfilt.R help four these)}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}
}
\value{
meta dataframe contianing flux calibrations
}
\description{
Function finds matches between standard stars on a current frame and an input
standard star catalogue. Calucultes photometric zero points for the fully reduced 
spliced data and each ccd individually. Also provides flux scaling between counts and
ergs/sec/cm^2/ang. Writes a meta data file will all information. Note that for the example to
work you must have unpacked the TAZ data files using:

> LibPaths<-.libPaths()

> system(paste('tar -xvf ', LibPaths, '/DEVILSTAZ/data/calibrators.tar --directory ', LibPaths, '/DEVILSTAZ/data/',sep=''))

> system(paste('tar -xvf ', LibPaths, '/DEVILSTAZ/data/idxFiles.tar --directory ', LibPaths, '/DEVILSTAZ/data/' ,sep=''))
}
\examples{
stdStars<-read.csv(paste(.libPaths(), '/DEVILSTAZ/data/calibrators/stdstars/stdStarCat.csv', sep=''))
fluxCal(file='object2df_red.fits', stdStars=stdStars, logName='tempLog.txt', verbose=1)
}
