% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.single.spec.R
\name{get.single.spec}
\alias{get.single.spec}
\title{Extracts single spectrum from FITS frame row}
\usage{
get.single.spec(file = file, hdu = 1, SpecNum = 1, tryMeta = FALSE,
  xunit = "Angstrom", yunit = "Counts", ID = NA, verbose = 2)
}
\arguments{
\item{file}{The file location}

\item{SpecNum}{FITS file row number}

\item{verbose}{0=quiet, 1=tell me some things that are going on,  2=tell me all things that are going on}
}
\description{
Extracts a spectrum from a given row in a FITS file
}
