% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileInfo.R
\name{getFileInfo}
\alias{getFileInfo}
\title{Get meta data information for the 2dF+AAOmega file.}
\usage{
getFileInfo(dir = dir, saveFile = saveFile, logName = logName,
  verbose = verbose)
}
\arguments{
\item{dir}{directory contianing files for which meta data should be retreived}

\item{saveFile}{filename to save meta data to}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}
}
\value{
meta dataframe for files in directoyr, dir.
}
\description{
Function searches through files in a given directory and 
returns a data table contining relavent meta data. Also saves meta data
to a given file.
}
\examples{
metaData<-getFileInfo(dir='data/reduced/run1_2017_12/2017_12_18/', saveFile='2017_12_18_metaData.Rdata', logName='tempLog.txt', verbose=1)
}
