% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfilt.R
\name{getfilt}
\alias{getfilt}
\title{Reads in filter responses for a specified filter.}
\usage{
getfilt(filter)
}
\arguments{
\item{filter}{string of the filter to load. Options are:
GALEX_FUV, GALEX_NUV, sdss_u,sdss_g,sdss_r,sdss_i,sdss_z, des_u,des_g,
des_r,des_i,des_z,des_Y,VISTA_Z,VISTA_Y,VISTA_J,VISTA_H,VISTA_K, WISE_W1,
WISE_W2,WISE_W3,WISE_W4,Herschel_100,Herschel_160,Herschel_250,Herschel_350,Herschel_500}
}
\value{
the filter response in the form c(rowNum, wavelength(ang), transmission)
}
\description{
Function reads in a specified filter tansmission curve from the DEVILSTAZ data directory. Note that for this to
work you must have unpacked the TAZ data files using:

> LibPaths<-.libPaths()

> system(paste('tar -xvf ', LibPaths, '/DEVILSTAZ/data/calibrators.tar --directory ', LibPaths, '/DEVILSTAZ/data/',sep=''))

> system(paste('tar -xvf ', LibPaths, '/DEVILSTAZ/data/idxFiles.tar --directory ', LibPaths, '/DEVILSTAZ/data/' ,sep=''))
}
\examples{
NUV_filt<-getfilt('GALEX_NUV')
u_filt<-getfilt('sdss_u')
r_filt<-getfilt('sdss_r')
Z_filt<-getfilt('VISTA_Z')
J_filt<-getfilt('VISTA_J')
K_filt<-getfilt('VISTA_K')

plot(r_filt[,2], r_filt[,3], type='l', col='green', xlab='Wavelength, Ang', ylab='Transmission', xlim=c(1000,25000), ylim=c(0,1))
lines(NUV_filt[,2], NUV_filt[,3], col='purple')
lines(u_filt[,2], u_filt[,3], col='cyan')
lines(Z_filt[,2], Z_filt[,3], col='orange')
lines(J_filt[,2], J_filt[,3], col='red')
lines(K_filt[,2], K_filt[,3], col='brown')
}
