% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanning.smooth.R
\name{hanning.smooth}
\alias{hanning.smooth}
\title{Hanning Smooth a Vector}
\usage{
hanning.smooth(arr1, degree = 3)
}
\arguments{
\item{arr1}{vector to smooth}

\item{degree}{degree of smoothing}
}
\value{
smoothed vector
}
\description{
Function to Hanning smooth a vector and return smoothed 
vector of the same size
}
\examples{
load(paste(.libPaths(),'/DEVILSTAZ/data/ExampleSpec.Rdata',sep=''))
SmoothSpec<-hanning.smooth(spec$flux, degree=17)
plot(spec$wave, spec$flux, type='l', xlab='Wavelength, ang', ylab='Counts')
lines(spec$wave,SmoothSpec, col='red')
}
