% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heliocentric.R
\name{Heliocentric}
\alias{Heliocentric}
\title{Calculate helocentric correction.}
\usage{
Heliocentric(ra, dec, epoch = 2000, jd = FALSE, tai = FALSE,
  longitude = 360 - 105.820417, latitude = 32.780361,
  altitude = 2788)
}
\arguments{
\item{ra}{Right ascension in degrees}

\item{dec}{Declination in degrees}

\item{epoch}{Epoch of observation for RA, DEC. Default is J2000.}

\item{jd}{Julian date}

\item{tai}{Number of seconds since Nov 17 1858}

\item{longitude}{Longitude of observatory}

\item{latitude}{Latitude of observatory}

\item{altitude}{altitude of observatory}
}
\value{
Velocity correction term, in km/s, to add to measured radial velocity 
to convert it to the heliocentric frame.
}
\description{
Function compute correction term to add to velocities to 
convert to heliocentric reference frame. Either jd or tai must be supplied.
}
\examples{
#Example for D10 field observed from the AAT on 18/12/2017.
vcorr<-Heliocentric(150.0, 1.55, epoch = 2000.0, tai = 58105.5, longitude =149.0661, latitude = -31.27704, altitude = 1164)
}
\author{
S. Burles

D. Schlegel

L. Davies (converted to R)
}
