% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.lines.R
\name{load.lines}
\alias{load.lines}
\title{Load common spectral feature list}
\usage{
load.lines()
}
\value{
dataframe of line names, wavelengths in various units, frequency and emission/absorption flag
}
\description{
Function to load common galaxy emission and absorption features.
}
\examples{
load(paste(.libPaths(),'/DEVILSTAZ/data/ExampleSpec.Rdata',sep=''))
#plot rest-frame spectrum
plot(spec$wave/(1+spec$z), hanning.smooth(spec$flux, degree=9), type='l', xlab='Wavelength, ang', ylab='Counts', ylim=c(-max(spec$flux,na.rm=T),max(spec$flux,na.rm=T)*1.1))
lineList<-load.lines()
print(lineList)
lines(c(lineList$wave_ang[2], lineList$wave_ang[2]), c(-max(spec$flux,na.rm=T),max(spec$flux,na.rm=T)*0.8), lty=2, col='blue')
text(lineList$wave_ang[2], max(spec$flux,na.rm=T),lineList$names[2], col='blue', cex=0.3)
lines(c(lineList$wave_ang[4], lineList$wave_ang[4]), c(-max(spec$flux,na.rm=T),max(spec$flux,na.rm=T)*0.8), lty=2, col='blue')
text(lineList$wave_ang[4], max(spec$flux,na.rm=T),lineList$names[4], col='blue', cex=0.3)
lines(c(lineList$wave_ang[5], lineList$wave_ang[5]), c(-max(spec$flux,na.rm=T),max(spec$flux,na.rm=T)*0.8), lty=2, col='blue')
text(lineList$wave_ang[5], max(spec$flux,na.rm=T),lineList$names[5], col='blue', cex=0.3)
lines(c(lineList$wave_ang[6], lineList$wave_ang[6]), c(-max(spec$flux,na.rm=T),max(spec$flux,na.rm=T)*0.8), lty=2, col='blue')
text(lineList$wave_ang[6], max(spec$flux,na.rm=T),lineList$names[6], col='blue', cex=0.3)
lines(c(lineList$wave_ang[12], lineList$wave_ang[12]), c(-max(spec$flux,na.rm=T),max(spec$flux,na.rm=T)*0.8), lty=2, col='darkgreen')
text(lineList$wave_ang[12], max(spec$flux,na.rm=T),lineList$names[12], col='darkgreen', cex=0.3)
lines(c(lineList$wave_ang[13], lineList$wave_ang[13]), c(-max(spec$flux,na.rm=T),max(spec$flux,na.rm=T)*0.8), lty=2, col='darkgreen')
text(lineList$wave_ang[13], max(spec$flux,na.rm=T),lineList$names[13], col='darkgreen', cex=0.3)
}
