% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.spec.R
\name{load.spec}
\alias{load.spec}
\title{Extracts spectra from AAT+AAOmega FITS frame}
\usage{
load.spec(file = file, SpecNums = 1, verbose = 2)
}
\arguments{
\item{file}{The file location of the AAOmega frame}

\item{SpecNums}{SpectraNumbers to Extract}

\item{verbose}{0=quiet, 1=tell me some things that are going on,  2=tell me all things that are going on}
}
\description{
Takes a reduced AAOOmegam frame and extracts spectra numbers given by
a vector, SpecNums. Produces a list with one object per spectrum containong all meta information 
wavelength and flux.
}
\examples{
allSpec<-load.spec(file="2018_1_15_config_6_reduced.fits", SpecNums=c(23,45,89), verbose=2)
names(allSpec)
plotSpec(spec=allSpec$spec23)
}
