% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magABspec.R
\name{magABspec}
\alias{magABspec}
\title{Measure Spectral AB magnitude in given filter}
\usage{
magABspec(spec, filter = "sdss_r")
}
\arguments{
\item{spec}{An R stucture containing spec$wave = vector of spectrum wavelengths
and spec$flux = vector of spectrum fluxes, spec$xunit = Units in the x-direction 
(Allowed units are 'ang', 'hz', 'm', 'micron','nm'), spec$yunit = Units in the y-direction 
(Allowed units are 'ang'=ergs/sec/cm^2/Ang, 'hz'=ergs/sec/cm^2/Hz, 'Jy'=Jansky.)}

\item{filter}{string of the filter to load.}
}
\value{
AB magnituide of spectrum in given filter
}
\description{
Convolves an input spectrum with a user-defined filter and calculate 
the spectral AB magnitude. Available filters are given in getfilt: GALEX_FUV, GALEX_NUV, 
sdss_u,sdss_g,sdss_r,sdss_i,sdss_z, des_u,des_g, des_r,des_i,des_z,des_Y,VISTA_Z,VISTA_Y,
VISTA_J,VISTA_H,VISTA_K, WISE_W1, WISE_W2,WISE_W3,WISE_W4,Herschel_100,Herschel_160,
Herschel_250,Herschel_350,Herschel_500. Note that for this to work you must have unpacked 
the TAZ data files using:

> LibPaths<-.libPaths()

> system(paste('tar -xvf ', LibPaths, '/DEVILSTAZ/data/calibrators.tar --directory ', LibPaths, '/DEVILSTAZ/data/',sep=''))

> system(paste('tar -xvf ', LibPaths, '/DEVILSTAZ/data/idxFiles.tar --directory ', LibPaths, '/DEVILSTAZ/data/' ,sep=''))
}
\examples{
load(paste(.libPaths(),'/DEVILSTAZ/data/ExampleSpec.Rdata',sep=''))

#scale to ergs/sec/cm^2/ang
spec$flux<-spec$flux*spec$fluxSc

plot(spec$wave, hanning.smooth(spec$flux, degree=9), type='l', xlab='Wavelength, ang', ylab='Counts')

filt<-getfilt('sdss_r')
lines(filt[,2],filt[,3]*max(spec$flux, na.rm=T), col='red')

print(magABspec(spec, filter='sdss_r'))
}
\author{
A. Robotham

L. Davies
}
