% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCalib.R
\name{makeCalib}
\alias{makeCalib}
\title{Make AAT master calibration files}
\usage{
makeCalib(toReduce = toReduce, doCalibQC = FALSE, logName = logName,
  verbose = verbose)
}
\arguments{
\item{toReduce}{directory path of raw data that needs to be reduced by TAZ in the DEVILS
data structure}

\item{doCalibQC}{TRUE/FALSE, run checkCal.R over the found calibration directory.}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}

\item{spec}{An R stucture containing spec$wave = vector of spectrum wavelengths
and spec$flux = vector of spectrum fluxes.}

\item{filter}{A filter with which to convolve with. A 2d matrix with column
1 as the filter wevelength (in the same units as spec$wave) and column
2 the filter response. Could be prodcued by \code{getfilt}.}
}
\description{
Reduce and combine all bias and dark frames associated with a givne directory. 
Function finds the run identifier in the DEVILS data structure that matches the current 
night to be reduced. Then identifies the corresponding bais and dark files. In no MASTER 
bias and/or dark is present, the function reduces all bias and dark frames and combines
to form a master bias and dark.
}
\examples{
makeCalib(toReduce='data/raw/run1_2017_12/2017_12_18/', doCalibQC=TRUE,logName='tempLog.txt', verbose=1)
}
