% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDOCats.R
\name{makeDOCats}
\alias{makeDOCats}
\title{Makes New DEVILS Observing Catalogues (DOCats)}
\usage{
makeDOCats(MASTERCat = MASTERCat, UserName = UserName,
  dateFor = dateFor, year = 2017, semester = "B", run = 1,
  logName = logName, verbose = verbose)
}
\arguments{
\item{MASTERCat}{current DEVILS MASTER catalogue to use}

\item{UserName}{your name to add to logs}

\item{dateFor}{the run and date these DOCats will be used for (e.g. 'run1_2017_12/2017_12_18')}

\item{year}{the current year (for SurveyInfo.txt)}

\item{semester}{the current semester (for SurveyInfo.txt)}

\item{run}{the current run number (for SurveyInfo.txt)}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}
}
\value{
New DOCats directory path
}
\description{
Function reads the current DEVILS Master Catlogue (DMCat), sky position catalogues,
standard star catalogues, guidestar catalogues and produces now DOCats for tiling. Also produces
a updated SurveyInfo.txt file required for the Tiling codes.
}
\examples{
makeDOCats(MASTERCat='data/catalogues/MASTERcats/DMCat2017-12-18.rda', UserName='NewUser', dateFor='run1_2017_12/2017_12_19', year=2017, semester='B', run=1, logName='tmpLog.txt', verbose=2)
}
