% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLog.R
\name{makeLog}
\alias{makeLog}
\title{Make Template Nightly Logs for DEVILS}
\usage{
makeLog(run = run, date = date, fileName = fileName)
}
\arguments{
\item{run}{the run date in the format run#_YYYY_MM}

\item{date}{the date in the format YYYY_MM_DD}

\item{fileName}{the output log filename}
}
\description{
Function produces template night logs for DEVILS obsering
for a given run and date.
}
\examples{
makeLog(run='run1_2017_12', date='2017_12_18',fileName='data/observing/run1_2017_12/2017_12_18/2017_12_18_DEVILS_obs_log.txt')
}
