% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftw.r
\name{planFFT}
\alias{planFFT}
\title{Glue between user and fftw.c}
\usage{
planFFT(n, effort = 0)
}
\description{
Functions for running fast Fourier Transform Routines in C.
}
\author{
Sebastian Krey <skrey@statistik.tu-dortmund.de>

Olaf Mersmann  <olafm@statistik.tu-dortmund.de>
}
