% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpec.basic.R
\name{plotSpec.basic}
\alias{plotSpec.basic}
\title{Plots a basic spectrum}
\usage{
plotSpec.basic(spec = spec, cex.axis = 1.4, cex.lab = 1.4,
  degSmooth = 7, xlim = NA, ylim = NA)
}
\arguments{
\item{spec}{spec object to plot}

\item{degSmooth}{Hanning smoothing to apply (odd number)}

\item{z}{FITS file row number}
}
\description{
Plots a basic spectrum with spec$flux and spec$wave paramters
}
