% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpec.R
\name{plotSpec}
\alias{plotSpec}
\title{Plots a spectrum object generated as part of TAZ}
\usage{
plotSpec(spec = spec)
}
\arguments{
\item{spec}{An R stucture containing spec$wave = vector of spectrum wavelengths
and spec$flux = vector of spectrum fluxes.}
}
\description{
Function takes a 'spec' R list with sepc$wave, spec$flux, ...
and plots the full 1D spectrum and zoom ins of key emission and absorption 
lines. Lines will be scaled to the correct redshfit in spec$z is provided. If 
no spec$z is provided, line positions and zooms are plotted in the rest-frame.
}
\examples{
load(paste(.libPaths(),'/DEVILSTAZ/data/ExampleSpec.Rdata',sep=''))
plotSpec(spec)
}
