% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessSpectrum.R
\name{ProcessSpectrum}
\alias{ProcessSpectrum}
\title{Internal AutoZ routines for processing spectra prior to fitting}
\usage{
ProcessSpectrum(specIn, stLambda = 3726, endLambda = 8850, os1 = 10,
  os2 = 60, minval = -10000, maxval = 1e+06, clipvalue = 25,
  baddataErrorValue = 1e+10, useInvCorrection = TRUE, verbose = TRUE)
}
\arguments{
\item{specIn}{input spectrum}

\item{stLambda}{min wavelength range to fit over}

\item{endLambda}{max wavelength range to fit over}

\item{os1}{length to extend below stLambda}

\item{os2}{length to extend above endLambda}

\item{minval}{min value to reject y values}

\item{maxval}{max value to reject y values}

\item{clipvalue}{value to itterative clip peaks}

\item{baddataErrorValue}{value to replace bad data with}

\item{useInvCorrection}{set relative flux calibration for GAMA}

\item{verbose}{TRUE/FALSE,  tell me whats going on}
}
\description{
A number of interal AutoZ routines.
}
\examples{
None applicable
}
\author{
I. Baldry
}
