% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCosmic.R
\name{RCosmic}
\alias{RCosmic}
\title{Cosmic Ray Rejection Functions, rewritten from PyCosmic (Copyright 2012 Bernd Husemann)}
\usage{
RCosmic(image, rdnoise = 1.8, sigma_det = 5, rlim = 1.2, iter = 5,
  fwhm_gauss = 2, gain = 1, verbose = 0)
}
\arguments{
\item{image}{image to find cosmics}
}
\description{
Cosmic ray rejection function. Detects and removes cosmics from astronomical 
images based on Laplacian edge detection scheme combined with a PSF convolution approach (Husemann  et al. in prep.).
}
\examples{

}
