% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readfiles.R
\name{RebinTempData}
\alias{RebinTempData}
\title{Internal AutoZ routines for reading in spectra and templates}
\usage{
RebinTempData(logLambda, templateNumbers = c(2:14, 16:22, 40:47),
  file = "../autozResouces/autoz-code/data/filtered-templates.fits",
  num_templates, verbose = TRUE)
}
\description{
A number of interal AutoZ routines.
}
\examples{
None applicable
}
\author{
I. Baldry
}
