% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run2dfDR.R
\name{run2dfDR}
\alias{run2dfDR}
\title{Run 2dFDR over a target directory}
\usage{
run2dfDR(toReduce = toReduce, doCalibQC = doCalibQC, logName = logName,
  verbose = verbose, reducCores = reducCores, doDark = T,
  doCosmic = doCosmic)
}
\arguments{
\item{toReduce}{directory path to reduce in the DEVILS data structure. Must be of the form, 
data/rawdata/run#_YYYY_MM/YYYY_MM_DD (i.e. data/rawdata/run1_2017_12/2017_12_18).}

\item{doCalibQC}{TRUE/FLASE, produce QC check plots for biases and darks}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}

\item{reducCores}{number of cores to use}

\item{doCosmic}{TRUE/FALSE do additonal cosmic ray rejection}
}
\description{
This is the highlevel main TAZ function for running 2dFDR (via aaorun) data reduction over 
a target directory. Function will find and reduce calibrators, make master bias and 
dark frames, produce QC plots, reduce arc, flat and object files, combine data files
for each ccd arm and splice to form final spectrum.
}
\examples{
run2dfDR(toReduce='data/rawdata/run1_2017_12/2017_12_18', doCalibQC=F, logName='tempLog.txt', verbose=1, reducCores=4)
}
