% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAutoZ.R
\name{runAutoZ}
\alias{runAutoZ}
\title{Run AutoZ over a supplied list of spectra}
\usage{
runAutoZ(specs = specs, logName = logName, verbose = verbose,
  makePlots = T, cores = cores, highZ = T)
}
\arguments{
\item{specs}{vector list of file paths to run AutoZ over. Must either be full directory path, or can
set to 'all' to run over all spectra in the 'data/reduced/stackedSpec/' directory.}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}

\item{makePlots}{TRUE/FLASE make plots of spectra using plotSpec.R}

\item{cores}{number of cores to use}
}
\description{
This is the highlevel main TAZ function for running AutoZ over a number of 
spectra, or all spectra in the 'data/reduced/stackedSpec/' directory. Function will run AutoZ
and make plots in the 'data/reduced/stackedSpec/AutoZplots' directory.
}
\examples{
runAutoZ(specs='all', logName='tempLog.txt', verbose=1, makePlots=T, cores=4)

runAutoZ(specs=c('data/reduced/stackedSpec/G007372.RData'), logName='tempLog.txt', verbose=1, makePlots=T, cores=4)
}
