% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUpLogLambda.R
\name{SetUpLogLambda}
\alias{SetUpLogLambda}
\title{Internal AutoZ function for setting up log wavelength scale}
\usage{
SetUpLogLambda(oversample = 5, lowValue = 3.3, highValue = 4,
  verbose = TRUE, highZ = T)
}
\arguments{
\item{oversample}{how much to oversample the spectrum given the wavelength gap 'sdssGap'}

\item{lowValue}{min values of scale}

\item{highValue}{max value of scale}

\item{verbose}{TRUE/FALSE, tell me what's going on}
}
\value{
logLambda wavelength scale
}
\description{
Set up log wavelength scale
}
\examples{
LamScale<-SetUpLogLambda(oversample=5, lowValue=3.3, highValue=4.0, verbose = TRUE)
}
\author{
I. Baldry
}
