% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackSpec.R
\name{stackSpec}
\alias{stackSpec}
\title{Stack repeat obervations with 2dF}
\usage{
stackSpec(ids = ids, logName = logName, verbose = verbose, makePlot = T,
  cores = cores)
}
\arguments{
\item{ids}{vector list of IDs to stack. Can be a list of IDs or set to 'all' which will stack all unique IDs in
the 'data/reduced/allSpec/' directory.}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}

\item{makePlot}{TRUE/FALSE, make diagnositc plots of each spectrum. These wil be written to data/reduced/stackedSpec/plots/}
}
\description{
This is the highlevel main TAZ function for stacking spectra. Function searches for 
repeat observations of a given ID in the DEVILS directory structure, and inverse variance weights
the full splice spectrum, each ccd arm individually, and the continuum extracted spectrum. Will also
produce diagnositc output plots for each spectrum to show individual spectra and stacks.
}
\examples{
stackSpec(ids='all',  logName='tempLog.txt', verbose=1,makePlots=T)
}
