% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateMASTERCat.R
\name{UpdateMASTERCat}
\alias{UpdateMASTERCat}
\title{Update DMcats with new redshifts and make DOCats}
\usage{
UpdateMASTERCat(cat = cat, specDir = specDir, OzDESGRC = OzDESGRC,
  logName = logName, verbose = verbose, makePlots = F,
  probGood = 0.9, useVIS = F)
}
\arguments{
\item{cat}{Path to current DMCat}

\item{specDir}{Path to directory containing R list spectra with value spec$z. 
Usually will be data/reduced/stackedSpec}

\item{OzDESGRC}{Either NA (do not update) or path to OzDES GRC cat to update to}

\item{logName}{log filename to write progress to}

\item{verbose}{tell me whats going on: 0=nothing, 1=somethings, 2=everything}

\item{makePlots}{make current survey progress plots}
}
\description{
This is the high-level main TAZ function for updating the DMCat with
new redshifts in a directory and generating DOCats for the 
next observing night.
}
\examples{

}
