% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VacuumFromAir.R
\name{VacuumFromAir}
\alias{VacuumFromAir}
\title{Internal AutoZ routines for converting vacuum wavelengths to air wavelengths}
\usage{
VacuumFromAir(air)
}
\arguments{
\item{air}{Wavelength in air}
}
\value{
Wavelength in vacuum
}
\description{
Function interpolate air wavelengths to vacuum wavelengths
using a lookup table. Equation comes from the SDSS web page:
http://www.sdss.org/dr7/products/spectra/vacwavelength.html. 
 Reference to Morton (1991, ApJS, 77, 119).
}
\examples{
Vac<-VacuumFromAir(3727.456)
}
\author{
I. Baldry
}
