% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCal.R
\name{checkCal}
\alias{checkCal}
\title{Produce QC diagnositic plots for dark and bias frames}
\usage{
checkCal(biasDir = biasDir, darksDir = darksDir, verbose = 1)
}
\arguments{
\item{biasDir}{directory containing bias frames}

\item{darksDir}{directory containing dark frames}

\item{verbose}{0 or 1, let me know whats going on. 0=no, 1=yes}
}
\description{
Function convolves a given spectrum with a filter band pass and returns either
convolved spectrum or a total sum within the filter. Note that for the example to
work you must have unpacked the TAZ data files using:
}
\examples{
checkCal(biasDir='data/biases/run1_2017_12/', darksDir='data/biases/run1_2017_12/', verbose=1)
}
